[FIX] record 내 필드 타입 수정 (#30)

+++ src/main/java/com/niedu/dto/course/StepListResponse.java
@@ -5,7 +5,7 @@ import com.niedu.dto.course.user_answer.AnswerResponse;
 import com.niedu.entity.course.StepType;
 
 public record StepListResponse (
-        Integer stepId,
+        Long stepId,
         Integer stepOrder,
         Boolean isCompleted,
         StepType contentType,
